+++ core/api/typescript/src/lib/enclaves/enclave_context.ts
@@ -30,7 +30,6 @@ import { GrpcWebApiContainerClient } from "./grpc_web_api_container_client";
 import type { GenericApiContainerClient } from "./generic_api_container_client";
 import { ModuleContext, ModuleID } from "../modules/module_context";
 import {
-    newDownloadFilesArtifactArgs,
     newGetModuleInfoArgs,
     newGetServiceInfoArgs,
     newLoadModuleArgs,
@@ -41,7 +40,7 @@ import {
     newRegisterServiceArgs,
     newRemoveServiceArgs,
     newRepartitionArgs,
-    newStartServiceArgs,
+    newStartServiceArgs, newStoreWebFilesArtifactArgs,
     newUnloadModuleArgs,
     newWaitForHttpGetEndpointAvailabilityArgs,
     newWaitForHttpPostEndpointAvailabilityArgs
@@ -544,14 +543,14 @@ export class EnclaveContext {
     }
 
     // Docs available at https://docs.kurtosistech.com/kurtosis-core/lib-documentation
-    public async downloadFiles(url: string): Promise<Result<FilesArtifactID, Error>> {
-        const args = newDownloadFilesArtifactArgs(url);
-        const downloadFilesArtifactResponseResult = await this.backend.downloadFilesArtifact(args)
-        if (downloadFilesArtifactResponseResult.isErr()) {
-            return err(downloadFilesArtifactResponseResult.error)
+    public async storeWebFiles(url: string): Promise<Result<FilesArtifactID, Error>> {
+        const args = newStoreWebFilesArtifactArgs(url);
+        const storeWebFilesArtifactResponseResult = await this.backend.storeWebFilesArtifact(args)
+        if (storeWebFilesArtifactResponseResult.isErr()) {
+            return err(storeWebFilesArtifactResponseResult.error)
         }
-        const downloadFilesArtifactResponse = downloadFilesArtifactResponseResult.value;
-        return ok(downloadFilesArtifactResponse.getUuid())
+        const storeWebFilesArtifactResponse = storeWebFilesArtifactResponseResult.value;
+        return ok(storeWebFilesArtifactResponse.getUuid())
     }
 
     // ====================================================================================================